presales_workflow/pyproject.toml
[project]
name = "presales-workflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Konrad Jarzyna",email = "konrad.jarzyna0@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "python-dotenv (>=1.0.0,<2.0.0)",
    "google-adk (>=0.5.0,<0.6.0)",
    "a2a-sdk (>=0.2.5,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src"

[tool.poetry.scripts]
start-web = "cli:start_web"
start-api = "cli:start_api"
start-dev = "cli:start_dev"


[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"
pylint = "^3.3.7"
black = "^25.1.0"
isort = "^6.0.1"

